Add explicit cast to non-apple platforms

+++ stdlib/public/Concurrency/DispatchGlobalExecutor.cpp
@@ -327,7 +327,7 @@ clock_and_value_to_time(int clock, long long sec, long long nsec) {
 #else
     struct timespec ts = { 
       .tv_sec = static_cast<long>(sec),
-      .tv_nsec = nsec
+      .tv_nsec = static_cast<long>(nsec)
     };
 #endif
     return dispatch_walltime(&ts, 0);
